+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Cat48: Add TrackQuality
 - Cat48: Add WarningErrorConditionsTargetClass
+- Cat48: Add Mode3ACodeConfidenceIndicator
 - Update Data Item docs
 
 ## [0.2.0] - 2020-08-21